requirements.txt
numpy
keras
sklearn
pillow
